// app/src/app/services/socket-event-types.ts
export enum SocketEventsTypes {
  CONNECTION = 'connection',
  CREATE_ROOM = 'createRoom',
  EDIT_ROOM = 'editRoom',
  DELETE_ROOM = 'deleteRoom',
  JOIN_ROOM = 'joinRoom',
  LEAVE_ROOM = 'leaveRoom',
  ROOM_DATA_UPDATE = 'roomDataUpdate',
  ROOMS_LIST_UPDATE = 'roomsListUpdate',
  KICK_USER_FROM_ROOM = 'kickUserFromRoom',
  BAN_USER_FROM_ROOM = 'banUserFromRoom',
  KICKED_FROM_ROOM = 'kickedFromRoom',
  BANNED_FROM_ROOM = 'bannedFromRoom',
  FETCH_ROOM = 'fetchRoom',
  FETCH_ALL_ROOMS = 'fetchAllRooms',
  FETCH_ALL_USER_ROOMS = 'fetchUserRooms',
  MESSAGE = 'message',
  SEND_MESSAGE = 'sendMessage',
  EDIT_MESSAGE = 'editMessage',
  DISCONNECT = 'disconnect',
  CONNECT_ERROR = 'connect_error',
}